Re-named card components, edit and open card modal component, create board modal component. Delete card function.

+++ server/controllers/closeBoardController.js
@@ -2,22 +2,30 @@ const Board = require('../models/Board');
 const Section = require('../models/Section');
 const Card = require('../models/Card');
 
-const closeBoardController = async (req, res) => {
-	try {
-		const closingBoard = req.body;
-
-		const deletingSections = await Section.find({ board: closingBoard.id });
+const cascadeDelete = async (closingBoard) => {
+	const deletingSections = await Section.find({ board: closingBoard.id });
+	if (deletingSections.length > 0) {
 		let deleteCardFilter = [];
 		deletingSections.forEach((section) => {
 			deleteCardFilter.push({ section: section._id });
 		});
 
-		// Deleting Cards that belongs to closing board
-		await Card.deleteMany({ $or: deleteCardFilter });
-		// Deleting Sections that belongs to closing board
-		await Section.deleteMany({ board: closingBoard.id });
-		// finally, Deleting/Closing boards
-		await Board.deleteMany({ _id: closingBoard.id });
+		if (deleteCardFilter.length > 0) {
+			// Deleting Cards that belongs to closing board
+			await Card.deleteMany({ $or: deleteCardFilter });
+		}
+	}
+
+	// Deleting Sections that belongs to closing board
+	await Section.deleteMany({ board: closingBoard.id });
+	// finally, Deleting/Closing boards
+	await Board.deleteMany({ _id: closingBoard.id });
+};
+
+const closeBoardController = async (req, res) => {
+	try {
+		const closingBoard = req.body;
+		cascadeDelete(closingBoard);
 
 		const selectedResponse = await Board.updateOne(
 			{},
